services/worker/requirements.txt
scrapy>=2.11.0
beautifulsoup4>=4.12.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
requests>=2.31.0